backend/requirements.txt
APScheduler==3.9.1
bcrypt==3.2.2 # Force Flask-Bcrypt to use a bcrypt version where the container doesn't need Rust to build (as there is no wheel for 4.0.2)
Flask-Bcrypt==1.0.1
beautifulsoup4==4.11.1
Flask==2.2.2
Flask-Session==0.4.0
PyJWT==2.6.0
python-dotenv==0.21.0
redis==4.3.4
redmail==0.4.2
requests==2.28.1
gunicorn==20.1.0
Flask-SQLAlchemy==3.0.2
